% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_runSum.R
\name{adapt_runSum}
\alias{adapt_runSum}
\title{Adaptive Running Sum}
\usage{
adapt_runSum(x, n = 1)
}
\arguments{
\item{x}{the given vector}

\item{n}{window size of the running sum: Default to 1. When n=1, returns x. When length(x)< n, return cumulative sum}
}
\description{
This function calculates the running sum of a given vector for a given window size.
}
\examples{
adapt_runSum(c(1, 2, 3, 4, 5))
}
\keyword{running}
\keyword{sum}
