% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course.R
\name{avg_table}
\alias{avg_table}
\title{Average Table
Knit the DFW table}
\usage{
avg_table(
  TYPE = "avg",
  display_table = by_course_DFW_table,
  analyzed_course = c(Analyzed_MATH_courses, advanced),
  num_sem_to_display = number_sem_display,
  FORM = "by_course",
  KEY = "DFW"
)
}
\arguments{
\item{TYPE}{can be "avg", "r-avg", etc}

\item{display_table}{the table to be knitted}

\item{analyzed_course}{courses to be displayed}

\item{num_sem_to_display}{number of semesters to be displayed}

\item{FORM}{can be "by_course" or others, see the code for details}

\item{KEY}{column name: such as "DFW", "total"}
}
\value{
knitted table avg_table
}
\description{
Average Table
Knit the DFW table
}
