% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_groups.R
\name{bargraph}
\alias{bargraph}
\title{Plot bargraph
plot the bargraph of the DFW of compared groups}
\usage{
bargraph(
  combined_db = DFW_dept_avg_Non_STEM_combined,
  earlist_sem = earliest_sem_in_plot,
  TYPE = "avg"
)
}
\arguments{
\item{combined_db}{combined database of all groups to be plotted}

\item{earlist_sem}{earliest semester to display}

\item{TYPE}{"avg" or "r_avg"}
}
\value{

}
\description{
Plot bargraph
plot the bargraph of the DFW of compared groups
}
