% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_groups.R
\name{bargraph_proportion}
\alias{bargraph_proportion}
\title{Bargraph of proportions}
\usage{
bargraph_proportion(
  combined_db = DFW_dept_avg_Non_STEM_combined,
  earlist_sem = earliest_sem_in_plot
)
}
\arguments{
\item{combined_db}{}

\item{earlist_sem}{}
}
\value{

}
\description{
Bargraph of proportions
}
