% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course.R
\name{barplot}
\alias{barplot}
\title{Plot barplot for top courses}
\usage{
barplot(DFW_db = by_course_DFW, TYPE = "avg")
}
\arguments{
\item{DFW_db}{}

\item{TYPE}{"avg" or "r_avg"}
}
\value{

}
\description{
Plot barplot for top courses
}
