% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course_college_major.R
\name{barplot_by_course_Major_DFW}
\alias{barplot_by_course_Major_DFW}
\title{Plot the barplots of top majors in a course for the past six semesters}
\usage{
barplot_by_course_Major_DFW(
  colleges = all_colleges,
  DFW_db = course_Major_DFW,
  course = 703,
  TYPE = "rDFW"
)
}
\arguments{
\item{colleges}{selected colleges}

\item{DFW_db}{DFW dataframe}

\item{course}{course number}

\item{TYPE}{"DFW" or "rDFW"}
}
\value{

}
\description{
Plot the barplots of top majors in a course for the past six semesters
}
