% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_number.R
\name{barplot_by_course_Major_total}
\alias{barplot_by_course_Major_total}
\title{Plot the barplot in the past six semesters for student totals in a course by major}
\usage{
barplot_by_course_Major_total(
  colleges = all_colleges,
  DFW_db = by_course_major_total,
  course = 703
)
}
\arguments{
\item{colleges}{selected colleges}

\item{DFW_db}{input database}

\item{course}{course number}
}
\value{

}
\description{
Plot the barplot in the past six semesters for student totals in a course by major
}
