% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course_college_major.R
\name{barplot_by_Major_best_DFW}
\alias{barplot_by_Major_best_DFW}
\title{barplot by Major of the past six semesters}
\usage{
barplot_by_Major_best_DFW(colleges = all_colleges, DFW_db = Major_DFW)
}
\arguments{
\item{colleges}{selected colleges}

\item{DFW_db}{DFW dataframe by Major}
}
\value{

}
\description{
barplot by Major of the past six semesters
}
