% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course_college_major.R
\name{barplot_by_Major_DFW}
\alias{barplot_by_Major_DFW}
\title{Major DFW barplot of the past six semesters
plot the DFW barplot of the past six semesters}
\usage{
barplot_by_Major_DFW(colleges = all_colleges, DFW_db = Major_DFW)
}
\arguments{
\item{colleges}{selected college}

\item{DFW_db}{DFW dataframe by Major}
}
\value{

}
\description{
Major DFW barplot of the past six semesters
plot the DFW barplot of the past six semesters
}
