% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_number.R
\name{barplot_by_Major_total}
\alias{barplot_by_Major_total}
\title{Plot the barplots of the past semesters for student totals by major}
\usage{
barplot_by_Major_total(colleges = all_colleges, DFW_db = by_major_total)
}
\arguments{
\item{colleges}{selected colleges}

\item{DFW_db}{by_major_total database}
}
\value{

}
\description{
Plot the barplots of the past semesters for student totals by major
}
