% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_faculty.R
\name{barplot_fac}
\alias{barplot_fac}
\title{Plot the bargraph of faculty DFW for the past six semesters with names}
\usage{
barplot_fac(DFW_db = by_faculty_DFW, TYPE = "avg")
}
\arguments{
\item{DFW_db}{input database}

\item{TYPE}{"avg" or "r_avg"}
}
\value{

}
\description{
Plot the bargraph of faculty DFW for the past six semesters with names
}
