% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course_college_major.R
\name{by_course_Major_running_DFW}
\alias{by_course_Major_running_DFW}
\title{Compute by course by Major running DFW}
\usage{
by_course_Major_running_DFW(db = by_course_by_Major, running_window = 1)
}
\arguments{
\item{db}{by_course_by_major dataframe}

\item{running_window}{running window size}
}
\value{

}
\description{
Compute by course by Major running DFW
}
