% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course_college_major.R
\name{college_running_DFW}
\alias{college_running_DFW}
\title{Ruunning DFW by collge}
\usage{
college_running_DFW(db = by_course_college, running_window = 1)
}
\arguments{
\item{db}{dataframe by college}

\item{running_window}{running window size}
}
\value{

}
\description{
Ruunning DFW by collge
}
