% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_groups.R
\name{combined_dept_avg_db}
\alias{combined_dept_avg_db}
\title{Calculate combined dataframe of groups and the group average
Calculate combined dataframe of a group of courses and the group average}
\usage{
combined_dept_avg_db(
  ls_db,
  gp_names = c("p7p8", "n7p8", "p9", "n789"),
  running_window = 1
)
}
\arguments{
\item{ls_db}{a list of all input dataframes}

\item{gp_names}{}

\item{running_window}{default to 1, not calculating the running average}
}
\value{

}
\description{
Calculate combined dataframe of groups and the group average
Calculate combined dataframe of a group of courses and the group average
}
