% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_faculty.R
\name{Dept_faculty_running_DFW}
\alias{Dept_faculty_running_DFW}
\title{Department running average of all MATH faculty (excluding CSCI faculty)}
\usage{
Dept_faculty_running_DFW(db = by_faculty, running_window = 1)
}
\arguments{
\item{db}{input database}

\item{running_window}{running window size}
}
\value{

}
\description{
Department running average of all MATH faculty (excluding CSCI faculty)
}
