% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course.R
\name{Dept_running_DFW}
\alias{Dept_running_DFW}
\title{Department running average
Department running average of only MATH courses (excluding ENGR)}
\usage{
Dept_running_DFW(db = by_course, running_window = 1)
}
\arguments{
\item{db}{input dataframe}

\item{running_window}{running window size. Default =1}
}
\value{
Dept_running_DFW
}
\description{
Department running average
Department running average of only MATH courses (excluding ENGR)
}
\examples{

}
