% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_groups.R
\name{Dept_running_DFW_seq_course}
\alias{Dept_running_DFW_seq_course}
\title{Department running DFW for a sequence of two courses}
\usage{
Dept_running_DFW_seq_course(
  db1 = df_700_703,
  db2 = df_p7_to_8,
  running_window = running_DFW_terms
)
}
\arguments{
\item{db1}{}

\item{db2}{}

\item{running_window}{running window size}
}
\value{

}
\description{
Department running DFW for a sequence of two courses
}
