% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course.R
\name{DFW_table_by_course}
\alias{DFW_table_by_course}
\title{DFW Table
Display DFW Table of selected courses with department average at the bottom}
\usage{
DFW_table_by_course(
  Dept_DFW = MATH_dept_running_DFW,
  by_course_DFW = by_course_running_DFW,
  TYPE = "r_avg",
  FORM = "by_course",
  KEY = "DFW"
)
}
\arguments{
\item{Dept_DFW}{Department DFW dataframe}

\item{by_course_DFW}{by_course_DFW data frame}

\item{TYPE}{can be "r_avg" for running average for "avg" for average}

\item{FORM}{can be "by_course" or some others}

\item{KEY}{column name: can be "DFW" or "total"}
}
\value{
DFW_table_by_course
}
\description{
DFW Table
Display DFW Table of selected courses with department average at the bottom
}
