% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_faculty.R
\name{DFW_table_by_faculty}
\alias{DFW_table_by_faculty}
\title{Generate the faculty DFW table}
\usage{
DFW_table_by_faculty(
  Dept_DFW = MATH_dept_faculty_running_DFW,
  by_faculty_DFW = by_faculty_running_DFW,
  TYPE = "Dept r_avg"
)
}
\arguments{
\item{Dept_DFW}{department aggregated DFW dataframe}

\item{by_faculty_DFW}{by faculty DFW dataframe}

\item{TYPE}{"Dept r_avg" or}
}
\value{

}
\description{
Generate the faculty DFW table
}
