% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course.R
\name{display_plot_by_course}
\alias{display_plot_by_course}
\title{Title
Display the DFW table and plot DFW of a group of courses}
\usage{
display_plot_by_course(
  DFW_data = by_course_DFW,
  avg_data = MATH_dept_avg,
  grp = remedial_courses,
  grp_name = "remedial",
  input_table = by_course_DFW_table,
  TYPE = "avg",
  earliest_sem = earliest_sem_in_plot
)
}
\arguments{
\item{DFW_data}{}

\item{avg_data}{}

\item{grp}{}

\item{grp_name}{}

\item{input_table}{the table in wide format to be knitted}

\item{TYPE}{"avg", or "r_avg"}

\item{earliest_sem}{earliest semester to display in the plot}
}
\value{

}
\description{
Title
Display the DFW table and plot DFW of a group of courses
}
