% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_number.R
\name{display_plot_total_by_course}
\alias{display_plot_total_by_course}
\title{display table and plot DFW of a group of courses}
\usage{
display_plot_total_by_course(
  DFW_data = by_course_DFW,
  avg_data = MATH_dept_total,
  grp = remedial_courses,
  grp_name = "remedial",
  input_table = by_course_total_table,
  TYPE = "total",
  earliest_sem = earliest_sem_in_plot,
  course_comp = FALSE
)
}
\arguments{
\item{DFW_data}{DFW data}

\item{avg_data}{Aggregated data}

\item{grp}{course numbers of the group}

\item{grp_name}{}

\item{input_table}{the table to be knitted}

\item{TYPE}{"total" or other}

\item{earliest_sem}{earlisest semester to display}

\item{course_comp}{compare courses in the group or not. Default=FALSE.}
}
\value{

}
\description{
display table and plot DFW of a group of courses
}
