% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_groups.R
\name{display_seq_table}
\alias{display_seq_table}
\title{Display the table for a sequence of two courses}
\usage{
display_seq_table(
  db1 = df_700_703,
  db2 = df_p7_to_8,
  TYPE = "avg",
  analyzed_courses = c("800/803"),
  num_sem_to_display = 16,
  FORM = "7to8",
  KEY = "DFW",
  running_window = running_DFW_terms
)
}
\arguments{
\item{db1}{}

\item{db2}{}

\item{TYPE}{"avg" or "r_avg"}

\item{analyzed_courses}{}

\item{num_sem_to_display}{in the table, 16 for a letter size portrait}

\item{FORM}{various form names, see the codes for details}

\item{KEY}{column name: "DFW", "total", etc}

\item{running_window}{running window size, default =15}
}
\value{

}
\description{
Display the table for a sequence of two courses
}
