% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_groups.R
\name{display_table}
\alias{display_table}
\title{Display  table of a group of courses
Calculate and display (knit)  a table of a group fo courses}
\usage{
display_table(
  df = df_Non_STEM_w_coreq,
  TYPE = "avg",
  analyzed_courses = Non_STEM,
  num_sem_to_display = 16,
  FORM = "coreq",
  KEY = "DFW",
  running_window = 1
)
}
\arguments{
\item{df}{input dataframe}

\item{TYPE}{"avg" or "r_avg"}

\item{analyzed_courses}{}

\item{num_sem_to_display}{number of semesters to display in the table, 16 for letter size portrait format}

\item{FORM}{string of the name of the table to be displayed in the table}

\item{KEY}{column name: "DFW", "total", etc.}

\item{running_window}{running window size}
}
\value{

}
\description{
Display  table of a group of courses
Calculate and display (knit)  a table of a group fo courses
}
