% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_faculty.R
\name{fac_avg_table}
\alias{fac_avg_table}
\title{Knit the faculty DFW average table}
\usage{
fac_avg_table(
  TYPE = "Dept avg",
  display_table = by_fac_table,
  analyzed_fac = Analyzed_fac,
  number_sem_to_display = number_sem_display
)
}
\arguments{
\item{TYPE}{"Dept avg" or other}

\item{display_table}{displayed table to be knitted}

\item{analyzed_fac}{selected fac}

\item{number_sem_to_display}{}
}
\value{

}
\description{
Knit the faculty DFW average table
}
