% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_faculty.R
\name{faculty_running_DFW}
\alias{faculty_running_DFW}
\title{Compute faculty running DFW}
\usage{
faculty_running_DFW(db = by_faculty, running_window = 1)
}
\arguments{
\item{db}{input database}

\item{running_window}{running window size;}
}
\value{

}
\description{
Compute faculty running DFW
}
