% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course_college_major.R
\name{Major_running_DFW}
\alias{Major_running_DFW}
\title{Major running DFW
calculate running DFW by Major}
\usage{
Major_running_DFW(db = by_course_Major, running_window = 1)
}
\arguments{
\item{db}{dataframe by course and by Major}

\item{running_window}{running window size}
}
\value{

}
\description{
Major running DFW
calculate running DFW by Major
}
