% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_groups.R
\name{piegraph_proportion}
\alias{piegraph_proportion}
\title{Plot the pie chart of proportions}
\usage{
piegraph_proportion(
  combined_db = DFW_dept_avg_Non_STEM_combined,
  earlist_sem = earliest_sem_in_plot,
  gtitle = ""
)
}
\arguments{
\item{combined_db}{}

\item{earlist_sem}{}

\item{gtitle}{the title of the graph}
}
\value{

}
\description{
Plot the pie chart of proportions
}
