% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_number.R
\name{plot_by_college_bargraph}
\alias{plot_by_college_bargraph}
\title{Bargraph of percentage of student numbers by college}
\usage{
plot_by_college_bargraph(
  combined_db = by_college_total,
  comp_gp = all_colleges,
  earliest_sem = earliest_sem_in_plot
)
}
\arguments{
\item{combined_db}{college total data frame}

\item{comp_gp}{selected colleges}

\item{earliest_sem}{earlist semester to display}
}
\value{

}
\description{
Bargraph of percentage of student numbers by college
}
