% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_number.R
\name{plot_by_college_linegraph}
\alias{plot_by_college_linegraph}
\title{Plot line graph of percentages for student numbers by college}
\usage{
plot_by_college_linegraph(
  combined_db = by_college_total,
  comp_gp = all_colleges,
  earliest_sem = earliest_sem_in_plot
)
}
\arguments{
\item{combined_db}{by_college_total dataframe}

\item{comp_gp}{selected colleges}

\item{earliest_sem}{earlist semester to display}
}
\value{

}
\description{
Plot line graph of percentages for student numbers by college
}
