% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course.R
\name{plot_course_DFW}
\alias{plot_course_DFW}
\title{Plot Course DFW}
\usage{
plot_course_DFW(
  by_course_DFW_df = by_course_DFW,
  MATH_dept_avg_df = MATH_dept_avg,
  MATH_dept_running_DFW_df = MATH_dept_running_DFW,
  earliest_semester_in_plot = earliest_sem_in_plot
)
}
\arguments{
\item{by_course_DFW_df}{by_course_DFW_df}

\item{MATH_dept_avg_df}{}

\item{MATH_dept_running_DFW_df}{}

\item{earliest_semester_in_plot}{}
}
\value{

}
\description{
Plot Course DFW
}
