% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course_college_major.R
\name{plot_DFW_by_college_linegraph}
\alias{plot_DFW_by_college_linegraph}
\title{Plot DFW linegraph by college}
\usage{
plot_DFW_by_college_linegraph(
  combined_db = by_college_total,
  comp_gp = all_colleges,
  earliest_sem = earliest_sem_in_plot,
  TYPE = "DFW"
)
}
\arguments{
\item{combined_db}{combined database}

\item{comp_gp}{set of colleges}

\item{earliest_sem}{earliest semester to display}

\item{TYPE}{"DFW" or "rDFW"}
}
\value{

}
\description{
Plot DFW linegraph by college
}
