% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course.R
\name{plot_DFW_by_group}
\alias{plot_DFW_by_group}
\title{Plot line graphs of DFW by group of courses, and saving the figure in a folder ./figs that needs to be created prior}
\usage{
plot_DFW_by_group(
  DFW_data = by_course_DFW,
  avg_data = MATH_dept_avg,
  grp = remedial_courses,
  grp_nm = "remedial",
  TYPE = "avg",
  earliest_sem = earliest_sem_in_plot
)
}
\arguments{
\item{DFW_data}{}

\item{avg_data}{average department data frame}

\item{grp}{course numbers of the group of courses. for example c(700,800)}

\item{grp_nm}{a string for the group name to shown in the figure title}

\item{TYPE}{can be "avg", "r_avg"}

\item{earliest_sem}{the earliest semester to show in the figure}
}
\value{

}
\description{
Plot line graphs of DFW by group of courses, and saving the figure in a folder ./figs that needs to be created prior
}
