% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_distribution.R
\name{plot_DFW_distribution}
\alias{plot_DFW_distribution}
\title{Plot DFW distribution curves vs. Term}
\usage{
plot_DFW_distribution(
  combined_db = by_course_grades_combined,
  comp_gp = c("703", "Dept"),
  comp_grade = c("D", "F", "W"),
  earliest_sem = earliest_sem_in_plot
)
}
\arguments{
\item{combined_db}{combined database of the course and the departmetn average}

\item{comp_gp}{compared gorup names, a vector.}

\item{comp_grade}{compared grades}

\item{earliest_sem}{earliest semester to display in the graph}
}
\value{

}
\description{
Plot DFW distribution curves vs. Term
}
