% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_faculty.R
\name{plot_fac_DFW_distribution}
\alias{plot_fac_DFW_distribution}
\title{Plot faculty DFW grade distribution line graph}
\usage{
plot_fac_DFW_distribution(
  combined_db = by_fac_grades_combined,
  comp_gp = c(individual, "Dept"),
  comp_grade = c("D", "F", "W"),
  earliest_sem = earliest_sem_in_plot
)
}
\arguments{
\item{combined_db}{combined dataframe of both the faculty and the departmetn avg}

\item{comp_gp}{compared groups}

\item{comp_grade}{selected grades to compare}

\item{earliest_sem}{earliest semester to display}
}
\value{

}
\description{
Plot faculty DFW grade distribution line graph
}
