% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_faculty.R
\name{plot_fac_grade_bargraph}
\alias{plot_fac_grade_bargraph}
\title{plot faculty grade distribution compared to department avg}
\usage{
plot_fac_grade_bargraph(
  combined_db = by_fac_grades_combined,
  comp_gp = c(individual, "Dept"),
  earliest_sem = earliest_sem_in_plot
)
}
\arguments{
\item{combined_db}{input data frame combining the faculty and the department average}

\item{comp_gp}{compared groups}

\item{earliest_sem}{earliest semester to display}
}
\value{

}
\description{
plot faculty grade distribution compared to department avg
}
