% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_faculty.R
\name{plot_fculty_DFW}
\alias{plot_fculty_DFW}
\title{Plot faculty DFW linegraph}
\usage{
plot_fculty_DFW(
  by_faculty_DFW_df = by_faculty_DFW,
  MATH_dept_fac_avg_df = MATH_dept_fac_avg,
  MATH_dept_fac_running_avg_df = MATH_dept_fac_running_avg,
  earlist_sem = earliest_sem_in_plot
)
}
\arguments{
\item{by_faculty_DFW_df}{}

\item{MATH_dept_fac_avg_df}{}

\item{MATH_dept_fac_running_avg_df}{}

\item{earlist_sem}{earlist semester to display}
}
\value{

}
\description{
Plot faculty DFW linegraph
}
