% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_distribution.R
\name{plot_grade_distrib}
\alias{plot_grade_distrib}
\title{Barplot of grade distribution in a course compared to the department average}
\usage{
plot_grade_distrib(
  combined_db = by_course_grades_combined,
  comp_gp = c("703", "Dept"),
  earliest_sem = earliest_sem_in_plot
)
}
\arguments{
\item{combined_db}{combined data base inlcuding the course data and department average data}

\item{comp_gp}{a vector of group names}

\item{earliest_sem}{earliest semester to display}
}
\value{

}
\description{
Barplot of grade distribution in a course compared to the department average
}
