% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_groups.R
\name{plot_line_compare_DFW}
\alias{plot_line_compare_DFW}
\title{Line graph of compared groups
Plot a line graph of compare groups}
\usage{
plot_line_compare_DFW(
  db = DFW_dept_avg_Non_STEM_combined,
  grp_nm = "Non-STEM",
  TYPE = "avg",
  FORM = "coreq",
  earlist_sem = earliest_sem_in_plot
)
}
\arguments{
\item{db}{combined database of all groups}

\item{grp_nm}{}

\item{TYPE}{"avg" or "r_avg"}

\item{FORM}{form name to be shown in the plot title}

\item{earlist_sem}{earliest semester to display in the figure}
}
\value{

}
\description{
Line graph of compared groups
Plot a line graph of compare groups
}
