% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_number.R
\name{plot_total_by_college_bargraph}
\alias{plot_total_by_college_bargraph}
\title{Plot bargraph of total students by college}
\usage{
plot_total_by_college_bargraph(
  combined_db = by_college_total,
  comp_gp = all_colleges,
  earliest_sem = earliest_sem_in_plot
)
}
\arguments{
\item{combined_db}{by_college_total dataframe}

\item{comp_gp}{selected colleges}

\item{earliest_sem}{earliest semester to display}
}
\value{

}
\description{
Plot bargraph of total students by college
}
