% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_number.R
\name{plot_total_by_group}
\alias{plot_total_by_group}
\title{plot the total number of students for a group of courses compared to department total}
\usage{
plot_total_by_group(
  DFW_data = by_course_DFW,
  avg_data = MATH_dept_total,
  grp = remedial_courses,
  grp_nm = "remedial",
  TYPE = "total",
  earliest_sem = earliest_sem_in_plot,
  course_comp = FALSE
)
}
\arguments{
\item{DFW_data}{by course DFW dataframe}

\item{avg_data}{Department aggregated dataframe}

\item{grp}{coure numbers of the group}

\item{grp_nm}{}

\item{TYPE}{"total" or "r_total"}

\item{earliest_sem}{earliest semester to display}

\item{course_comp}{compare courses in the group or not. Default=FALSE.}
}
\value{

}
\description{
plot the total number of students for a group of courses compared to department total
}
