% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course.R
\name{running_DFW}
\alias{running_DFW}
\title{Running DFW
Calculating running DFW, when windowsize=1, it just calculates averages}
\usage{
running_DFW(db = by_course, running_window = 1)
}
\arguments{
\item{db}{input dataframe}

\item{running_window}{running window size. Default to 1.}
}
\value{
by_course_running_DFW
}
\description{
Running DFW
Calculating running DFW, when windowsize=1, it just calculates averages
}
\examples{

}
