% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course.R
\name{table_by_course}
\alias{table_by_course}
\title{Knit the DFW Table
Knit the DFW table by course}
\usage{
table_by_course(ls, number_sem_to_display = number_sem_display, TYPE = "avg")
}
\arguments{
\item{ls}{a list contains the DFW table to be knitted. The table is given by ls$display}

\item{number_sem_to_display}{in the table}

\item{TYPE}{"avg" or "r_avg"}
}
\value{

}
\description{
Knit the DFW Table
Knit the DFW table by course
}
