% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_faculty.R
\name{table_individual}
\alias{table_individual}
\title{knit the table for individual DFW}
\usage{
table_individual(i = 1, list, TYPE = "avg")
}
\arguments{
\item{i}{index for the individua in the group}

\item{list}{list that contains the displayed table}

\item{TYPE}{"avg" or "r_avg"}
}
\value{

}
\description{
knit the table for individual DFW
}
