% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course_college_major.R
\name{top_best_Major_by_DFW}
\alias{top_best_Major_by_DFW}
\title{top 20 best majors for a semeser
plot the top 20 majors for a semester in a college (a group of colleges)}
\usage{
top_best_Major_by_DFW(
  colleges = all_colleges,
  term = current_term,
  DFW_db = Major_DFW,
  TYPE = "rDFW"
)
}
\arguments{
\item{colleges}{selected colleges}

\item{term}{the term to plot the bargraph}

\item{DFW_db}{DFW dataframe by Major}

\item{TYPE}{"rDFW" or "DFW"}
}
\value{

}
\description{
top 20 best majors for a semeser
plot the top 20 majors for a semester in a college (a group of colleges)
}
