% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course.R
\name{top_DFW_courses}
\alias{top_DFW_courses}
\title{Plot top DFW courses bargraph}
\usage{
top_DFW_courses(TYPE = "avg", term = current_term, DFW_db = by_course_DFW)
}
\arguments{
\item{TYPE}{can be "avg", "r_avg"}

\item{term}{semester code, such as 202102}

\item{DFW_db}{DFW dataframe}
}
\value{

}
\description{
Plot top DFW courses bargraph
}
