% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_faculty.R
\name{top_DFW_faculty}
\alias{top_DFW_faculty}
\title{plot faculty DFW bargraph with names}
\usage{
top_DFW_faculty(TYPE = "avg", term = current_term, DFW_db = by_faculty_DFW)
}
\arguments{
\item{TYPE}{"avg" or "r_avg"}

\item{term}{term to display}

\item{DFW_db}{input data frame}
}
\value{

}
\description{
plot faculty DFW bargraph with names
}
