% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_faculty.R
\name{top_DFW_faculty_no_name}
\alias{top_DFW_faculty_no_name}
\title{Plot bargraph of top 20 faculty without showing names}
\usage{
top_DFW_faculty_no_name(term = current_term)
}
\arguments{
\item{term}{term to plot}
}
\value{

}
\description{
Plot bargraph of top 20 faculty without showing names
}
