% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course_college_major.R
\name{top_Major_by_course_DFW}
\alias{top_Major_by_course_DFW}
\title{Compute the top 20 Majors in a course}
\usage{
top_Major_by_course_DFW(
  colleges = all_colleges,
  term = current_term,
  DFW_db = course_Major_DFW,
  course = 703,
  TYPE = "rDFW"
)
}
\arguments{
\item{colleges}{selected colleges}

\item{term}{term for the computation}

\item{DFW_db}{DFW data base by course and Major}

\item{course}{course number}

\item{TYPE}{"rDFW" or "DFW"}
}
\value{

}
\description{
Compute the top 20 Majors in a course
}
