% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_number.R
\name{top_Major_by_course_total}
\alias{top_Major_by_course_total}
\title{bargraph of top 20 majors of student numbers in a course}
\usage{
top_Major_by_course_total(
  colleges = all_colleges,
  term = current_term,
  DFW_db = by_course_major_total,
  course = 703
)
}
\arguments{
\item{colleges}{selected colleges}

\item{term}{term for which the graph is generated}

\item{DFW_db}{input database}

\item{course}{course number}
}
\value{

}
\description{
bargraph of top 20 majors of student numbers in a course
}
