% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_course_college_major.R
\name{top_Major_by_DFW}
\alias{top_Major_by_DFW}
\title{top DFW majors (bargraph)
plot the bargraph of top 20 Majors of DFW for a college or all colleges}
\usage{
top_Major_by_DFW(
  colleges = all_colleges,
  term = current_term,
  DFW_db = Major_DFW,
  TYPE = "rDFW"
)
}
\arguments{
\item{colleges}{selected colleges}

\item{term}{the term to plot the bargraph}

\item{DFW_db}{Major DFW dataframe}

\item{TYPE}{"DFW" or "rDFW"}
}
\value{

}
\description{
top DFW majors (bargraph)
plot the bargraph of top 20 Majors of DFW for a college or all colleges
}
