% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_number.R
\name{top_Major_by_total}
\alias{top_Major_by_total}
\title{Bargraph of student total of top 20 Majors}
\usage{
top_Major_by_total(
  colleges = all_colleges,
  term = current_term,
  DFW_db = by_major_total
)
}
\arguments{
\item{colleges}{selected colleges}

\item{term}{the term for which the graph is created}

\item{DFW_db}{by_major_total data base}
}
\value{

}
\description{
Bargraph of student total of top 20 Majors
}
