% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_number.R
\name{total_table_by_course}
\alias{total_table_by_course}
\title{Knit the total table by course}
\usage{
total_table_by_course(
  ls,
  number_sem_to_display = number_sem_display,
  TYPE = "total"
)
}
\arguments{
\item{ls}{a list contains the table in ls$display}

\item{number_sem_to_display}{number of semester to display}

\item{TYPE}{"total", "DFW" or others}
}
\value{

}
\description{
Knit the total table by course
}
